Deleted hipchat_notifier.py, Added slack_notifier.py and Integrated slack into shippable.yml
