+++ shippable.yml
@@ -3,8 +3,8 @@ go:
     - 1.4
 env:
     global:
-        - PROJECT=Foundation HIPCHAT_ROOM=Foundation
-        - secure: JePvX8e49vPPWRLagd9lvuwV8wRBUD1EDkHnANJ8bTzfnDiAPBgqrgocliz7KR6F5ZPHBAvRkrIa6ZP1PVLdC/LeOzoojFyPzU7iiKxkatT2+Ry+QKnH603UGJzIfui1bLQ/eZT2D8wOjMggs4z8DXH62HCmzl+ovhboU1XBYRBuYVMZBKKN/2kcz8LVTZ0BP9jLLMJ5xHb9PDOX7fwBREnjn0dN2yy6uNpz9t7rKJCxNf4uk7+M6AHrryU4mEHpEfu9wbUhMGve5Del78YcZfc5PMDtmKqvamUDICOMCzHRFdl3kpkNsr72ynoZ3s0Vus/3cH9hShCfHNacVk4xQA==
+        - SLACK_ORG=Ottemo PROJECT=Foundation
+        - secure: IgdstVqJgSEKc/DjAhAiySFT43+GFxnm5Vejtb2fToZtMO+D1CFxn3tmFk3DOlHqBlBQPm6ADgW9lUbg/x1G7Kn2RyfMvbV2AkzRJzqfaHc5lXfUw+d5dniVyqcet/nC3Lul5C2fEsuHLsB1dTKunzGXJowHMTU/ZG4BdcgA9zM6h7XuGWO6euJylDpNfZ9nz9V9r14Wbxx2ndeaXgVtWXNpSdu85R1DaAKRzmuObtYmsnoKS5F6mL+v6UuM2A/4q08DRtD52jZwavrVOksAZynmLgjm2/wB2Ew6lZgPlRfCGdzRzCgIP5RenJd7jfrW2Q9tNrOGPpP8BEGjjBg7KQ==
 before_install:
     - shippable_retry sudo apt-get update
     - shippable_retry sudo apt-get install -y bzr 
@@ -23,8 +23,8 @@ script:
     - go test -v ./...
 
 after_failure:
-    - python hipchat_notifier.py --project $PROJECT --room $HIPCHAT_ROOM --token $HIPCHAT_TOKEN
+    - python slack_notifier.py --org $SLACK_ORG --project $PROJECT --token $SLACK_TOKEN
 
 after_success:
-    - python hipchat_notifier.py --project $PROJECT --room $HIPCHAT_ROOM --token $HIPCHAT_TOKEN -s
+    - python slack_notifier.py --org $SLACK_ORG --project $PROJECT --token $SLACK_TOKEN -s
